src/base/requirements.txt
absl-py==2.2.2
albucore==0.0.24
albumentations==2.0.8
alembic==1.15.2
annotated-types==0.7.0
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
babel==2.14.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.9.0
Bottleneck==1.3.8
Brotli==1.1.0
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.2.1
cloudpickle==3.0.0
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
databricks-sdk==0.53.0
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.18
docker==7.1.0
executing==2.0.1
fastapi==0.115.12
fastjsonschema==2.19.1
filelock==3.14.0
Flask==3.1.1
fonttools==4.52.4
fsspec==2024.6.0
gitdb==4.0.12
GitPython==3.1.44
gmpy2==2.1.5
google-auth==2.40.1
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.0.3
grpcio==1.62.2
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib_metadata==8.6.1
ipykernel==6.29.4
ipython==8.24.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.25
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.2
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.14.1
jupyter_server_terminals==0.5.3
jupyterlab==4.2.1
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.2
kiwisolver==1.4.5
Mako==1.3.10
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.10.0
matplotlib-inline==0.1.7
mistune==3.0.2
mkl-service==2.4.2
mkl_fft==1.3.14
mkl_random==1.2.10
mlflow==2.22.0
mlflow-skinny==2.22.0
mpmath==1.3.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
notebook_shim==0.2.4
numexpr==2.10.0
numpy==2.2.5
opencv-python-headless==4.11.0.86
opentelemetry-api==1.33.1
opentelemetry-sdk==1.33.1
opentelemetry-semantic-conventions==0.54b1
overrides==7.7.0
packaging==24.0
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.2
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==6.31.0
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.18.0
pyparsing==3.1.2
PyQt6==6.7.1
PyQt6_sip==13.6.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.35.1
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.18.1
rsa==4.9.1
scikit-learn==1.5.0
scipy==1.15.3
seaborn==0.13.2
Send2Trash==1.8.3
setuptools==72.1.0
simsimd==6.2.1
sip==6.8.3
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.41
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.46.2
stringzilla==3.12.5
sympy==1.12
tensorboard==2.19.0
tensorboard-data-server==0.7.2
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.3.0
torch==2.7.1
torchaudio==2.7.0+cpu
torchvision==0.22.0+cpu
tornado==6.4
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2024.1
unicodedata2==15.1.0
urllib3==2.2.2
uvicorn==0.34.2
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
zipp==3.18.1